Allow sequential jobs to continue by default
It's likely that since jobs ran in parallel before, there were no dependencies between jobs. Therefore, now that sequential jobs are possible, we should continue execution of the job sequence even if one job fails. To override this and halt the sequence on failure, individual jobs can set `halt_sequence_on_failure: true` to suffix their command with an "&&" instead of a ";".